perf: refactor wizard to avoid expensive imports at CLI startup
